requirements.txt
altgraph==0.17
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2020.11.8
chardet==3.0.4
cx-Freeze==6.3
Django==3.0.5
django-debug-toolbar==3.2
future==0.18.2
idna==2.10
importlib-metadata==2.0.0
pefile==2019.4.18
Pillow==8.3.1
psycopg2==2.8.6
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
python-memcached==1.59
pytz==2019.3
pywin32-ctypes==0.2.0
requests==2.25.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.26.2
zipp==3.4.0
